requirements.txt
chromadb
openai
python-dotenv
pytest
sentence-transformers
